PR-E-10: extract StateMachineEmitter (async/iterator synthesis + MoveNext/SetStateMachine IL)
Splits async/iterator state-machine synthesis and the top-level MoveNext /
SetStateMachine / AwaitOnCompleted / AsyncKickoff IL emission out of the
13k-line ReflectionMetadataEmitter into a new internal sealed class
StateMachineEmitter at src/Core/CodeAnalysis/Emit/StateMachineEmitter.cs.
Follows the same Option-B template as PR-E-9 (ClosureEmitter) and PR-E-5
(ConversionEmitter).

Option chosen: B
-----------------
The 2 BodyEmitter-internal state-machine methods
(EmitStateMachineAwaitOnCompleted, EmitAsyncIteratorBuilderMoveNext) stay
on BodyEmitter and are deferred to PR-E-11 MethodBodyEmitter. They run
inside BodyEmitter's active method-emit session and reach into private
state (il, locals, parameters, currentPlan, asyncIteratorEmitCtx) that
would require widening BodyEmitter's surface to move now. Consistent with
the same decision in PR-E-9.

For the BodyEmitter-driven portion of EmitStateMachineMoveNext (the
locals/labels pre-scan + BodyEmitter run that produces the raw IL bytes),
a small private helper BuildMoveNextBodyBytes is added on
ReflectionMetadataEmitter and exposed to StateMachineEmitter via an
injected Func<AsyncStateMachinePlan, MoveNextBodyResult> — the same
callback pattern PR-E-8 TypeDefEmitter uses for static-ctor body bytes.
This helper will move into BodyEmitter / MethodBodyEmitter in PR-E-11.

What moves into StateMachineEmitter
-----------------------------------
Nested classes (previously private nested on RME):
- IteratorStateMachineInfo   (was ~6955)
- AsyncIteratorEmitContext   (was ~6972)

Plus a new public readonly struct MoveNextBodyResult that carries the
five values returned by BuildMoveNextBodyBytes.

Top-level orchestration / build methods:
- EmitAsyncKickoffBody                    (was ~3286)
- SynthesizeIteratorStateMachines         (was ~4790)
- SynthesizeAsyncIteratorStateMachines    (was ~4911)
- BuildMoveNextAsyncBody                  (was ~5104)
- BuildDisposeAsyncBody                   (was ~5173)
- BuildGetAsyncEnumeratorBody             (was ~5219)
- BuildVtsGetStatusBody                   (was ~5312)
- BuildVtsGetResultBody                   (was ~5329)
- BuildVtsOnCompletedBody                 (was ~5346)
- BuildVtsGetResultVoidBody               (was ~5367)
- CreateIteratorStateMachineLiteral       (was ~4884)
- CreateAsyncIteratorKickoffLiteral       (was ~5387)
- SynthesizeAsyncLambdaStateMachines      (was ~5446)
- GetStateMachineStartMethodSpec          (was ~3386)

Top-level state-machine IL emit methods:
- EmitStateMachineMoveNext         (was ~3111)
- EmitStateMachineSetStateMachine  (was ~3248)
- EmitAwaitOnCompletedCall         (was ~3409)

State-machine caches / state:
- IteratorKickoffBodies        (was iteratorKickoffBodies ~106)
- IteratorStateMachineInfos    (was iteratorStateMachineInfos ~107)
- AsyncStateMachinePlans       (was asyncStateMachinePlans ~110, with internal setter to preserve the assignment from Emit())
- IteratorPlans                (was iteratorPlans ~113, with internal setter)
- AsyncIteratorPlans           (was asyncIteratorPlans ~116, with internal setter)
- AsyncIteratorInfos           (was asyncIteratorInfos ~121)
- AsyncIteratorEmitContexts    (was asyncIteratorEmitContexts ~124)
- AsyncSmEnclosingClosures     (was asyncSmEnclosingClosures ~128)

What stays on ReflectionMetadataEmitter
---------------------------------------
- lambdaBodies (cross-cutting; populated by both ClosureEmitter and
  StateMachineEmitter, read by EmitFunction). RME passes a shared
  reference to StateMachineEmitter's ctor — same pattern as PR-E-9.
- AddIteratorInterfaceImplementations and
  AddAsyncIteratorInterfaceImplementations consume RME-internal services
  (cache, GetTypeReference, GetTypeHandleForMember) and stay on RME.
  They now accept the moved nested classes via qualified names
  (StateMachineEmitter.IteratorStateMachineInfo etc.).
- BuildMoveNextBodyBytes — the BodyEmitter-driven body-bytes helper
  described above, deferred to PR-E-11.
- The 2 BodyEmitter-internal methods listed above (Option B).

Wiring
------
- ReflectionMetadataEmitter holds the new collaborator as
  private StateMachineEmitter stateMachines and instantiates it inside
  EmitCore once EmitContext + collaborators are wired.
- Every former this.X / this.outer.X reference to a moved member now
  reads from this.stateMachines.X (and the few BodyEmitter call sites
  read from this.outer.stateMachines.X).
- StateMachineEmitter has zero direct dependency on
  ReflectionMetadataEmitter — composition + 7 injected callbacks (type
  reference / handle / method-spec accessors, NextParameterHandle,
  EncodeClrType, and BuildMoveNextBodyBytes) per the established pattern.

Determinism / IL identity
-------------------------
- RefactoringBaselineTests: zero hash diffs (verified locally against
  test/Core.Tests/Baselines/refactoring-baseline.json — 97/98 pass with
  the pre-existing RegenerateBaseline skip).
- DeterministicEmitTests: green.
- SlotDictionaryAliasingAssertionTests: green.
- All Core.Tests pass (2120/2121 with the same RegenerateBaseline skip).
- Compiler.Tests cannot be fully exercised locally on macOS due to the
  256-fd default rlimit (which causes 'Too many open files' on both
  this branch and clean main); CI will exercise the full matrix.

LoC delta
---------
- StateMachineEmitter.cs:         1283 (new)
- ReflectionMetadataEmitter.cs:  -928 net (1232 deleted, 304 added for
  the new field, wiring, and BuildMoveNextBodyBytes helper)

Next PR in the sequence: E-11 MethodBodyEmitter — the big one that
absorbs the BodyEmitter-internal methods deferred from E-5, E-9, E-10
(EmitConversion family, all closure body-emit methods, the 2 SM body-
emit methods, plus BuildMoveNextBodyBytes itself).
